refactor namespace for clarity
