+++ src/openapi2excel.tests/OpenApiDocumentationGeneratorTest.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Xml.Linq;
 using openapi2excel.core;
 using openapi2excel.core.Builders;
-using openapi2excel.core.Builders.CustomXml;
+using openapi2excel.core.Builders.CommentsManagement;
 using Microsoft.OpenApi.Readers;
 using ClosedXML.Excel;
 using openapi2excel.core.Common;
