+++ src/openapi2excel/Builders/CommentsManagement/ThreadedCommentWithContext.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Person = DocumentFormat.OpenXml.Office2019.Excel.ThreadedComments.Person;
 
-namespace openapi2excel.core.Builders.CustomXml;
+namespace openapi2excel.core.Builders.CommentsManagement;
 
 /// <summary>
 /// Data structure to capture threaded comment with its worksheet context and OpenAPI mapping information.
