+++ src/openapi2excel/Builders/WorksheetPartsBuilders/Common/OpenApiSchemaDescriptor.cs
@@ -1,6 +1,6 @@
 using ClosedXML.Excel;
 using Microsoft.OpenApi.Models;
-using openapi2excel.core.Builders.CustomXml;
+using openapi2excel.core.Builders.CommentsManagement;
 using openapi2excel.core.Common;
 
 namespace openapi2excel.core.Builders.WorksheetPartsBuilders.Common;
